VPN-4009: Add configuration event streaming.
- Introduce `ConfigChanged` event in `TunnelEvent` to notify clients of
  configuration changes.
- `VpnServiceConfig` has moved to the `nym-vpn-lib-types` crate.
- Conversions from `VpnServiceConfig`, `EntryPoint`, and `ExitPoint` to
  their protobuf equivalents are infallible so their `impl TryFrom` is
  now `impl From`.
