+++ nym-vpn-core/crates/nym-vpn-lib-types-uniffi/src/lib.rs
@@ -11,6 +11,7 @@ mod gateway_directory;
 mod ip_pair;
 mod ipnetwork;
 mod network_config;
+mod service;
 mod std;
 mod time;
 mod tunnel_state_machine;
@@ -25,6 +26,7 @@ pub use gateway_directory::*;
 pub use ip_pair::*;
 pub use ipnetwork::*;
 pub use network_config::*;
+pub use service::*;
 pub use std::*;
 pub use time::*;
 pub use tunnel_state_machine::*;
