+++ nym-vpn-core/crates/nym-vpn-lib-types-uniffi/src/service.rs
@@ -1,0 +1,60 @@
+use crate::{EntryPoint, ExitPoint};
+
+#[derive(uniffi::Record)]
+pub struct VpnServiceConfig {
+    pub entry_point: EntryPoint,
+    pub exit_point: ExitPoint,
+    pub dns: Option<String>,
+    pub disable_ipv6: bool,
+    pub enable_two_hop: bool,
+    pub netstack: bool,
+    pub disable_poisson_rate: bool,
+    pub disable_background_cover_traffic: bool,
+    pub min_mixnode_performance: Option<u8>,
+    pub min_gateway_mixnet_performance: Option<u8>,
+    pub min_gateway_vpn_performance: Option<u8>,
+}
+
+impl From<nym_vpn_lib_types::service::VpnServiceConfig> for VpnServiceConfig {
+    fn from(value: nym_vpn_lib_types::service::VpnServiceConfig) -> Self {
+        Self {
+            entry_point: value.entry_point.into(),
+            exit_point: value.exit_point.into(),
+            dns: value.dns.map(|ip| ip.to_string()),
+            disable_ipv6: value.disable_ipv6,
+            enable_two_hop: value.enable_two_hop,
+            netstack: value.netstack,
+            disable_poisson_rate: value.disable_poisson_rate,
+            disable_background_cover_traffic: value.disable_background_cover_traffic,
+            min_mixnode_performance: value.min_mixnode_performance,
+            min_gateway_mixnet_performance: value.min_gateway_mixnet_performance,
+            min_gateway_vpn_performance: value.min_gateway_vpn_performance,
+        }
+    }
+}
+
+impl From<VpnServiceConfig> for nym_vpn_lib_types::service::VpnServiceConfig {
+    fn from(value: VpnServiceConfig) -> Self {
+        Self {
+            entry_point: value.entry_point.into(),
+            exit_point: value.exit_point.into(),
+            dns: value.dns.and_then(|ip| ip.parse().ok()),
+            disable_ipv6: value.disable_ipv6,
+            enable_two_hop: value.enable_two_hop,
+            netstack: value.netstack,
+            disable_poisson_rate: value.disable_poisson_rate,
+            disable_background_cover_traffic: value.disable_background_cover_traffic,
+            min_mixnode_performance: value.min_mixnode_performance,
+            min_gateway_mixnet_performance: value.min_gateway_mixnet_performance,
+            min_gateway_vpn_performance: value.min_gateway_vpn_performance,
+        }
+    }
+}
+
+pub type BoxedVpnServiceConfig = Box<VpnServiceConfig>;
+
+uniffi::custom_type!(BoxedVpnServiceConfig, VpnServiceConfig, {
+    remote,
+    try_lift: |val| Ok(Box::new(val)),
+    lower: |val| *val
+});
