+++ nym-vpn-core/crates/nym-vpn-lib-types/src/service.rs
@@ -1,0 +1,49 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use std::{fmt, net::IpAddr};
+
+use nym_gateway_directory::{EntryPoint, ExitPoint};
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct VpnServiceConfig {
+    pub entry_point: EntryPoint,
+    pub exit_point: ExitPoint,
+    pub dns: Option<IpAddr>,
+    pub disable_ipv6: bool,
+    pub enable_two_hop: bool,
+    pub netstack: bool,
+    pub disable_poisson_rate: bool,
+    pub disable_background_cover_traffic: bool,
+    pub min_mixnode_performance: Option<u8>,
+    pub min_gateway_mixnet_performance: Option<u8>,
+    pub min_gateway_vpn_performance: Option<u8>,
+}
+
+impl fmt::Display for VpnServiceConfig {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "entry point: {}, exit point: {}",
+            self.entry_point, self.exit_point
+        )
+    }
+}
+
+impl Default for VpnServiceConfig {
+    fn default() -> Self {
+        Self {
+            entry_point: EntryPoint::Random,
+            exit_point: ExitPoint::Random,
+            dns: None,
+            disable_ipv6: false,
+            enable_two_hop: false,
+            netstack: false,
+            disable_poisson_rate: false,
+            disable_background_cover_traffic: false,
+            min_mixnode_performance: None,
+            min_gateway_mixnet_performance: None,
+            min_gateway_vpn_performance: None,
+        }
+    }
+}
