+++ nym-vpn-core/crates/nym-vpn-lib-types/src/tunnel_event.rs
@@ -5,6 +5,8 @@ use std::fmt;
 
 use si_scale::helpers::bibytes2;
 
+use crate::service::VpnServiceConfig;
+
 #[cfg(feature = "nym-type-conversions")]
 use nym_bandwidth_controller::BandwidthStatusMessage;
 #[cfg(feature = "nym-type-conversions")]
@@ -20,6 +22,7 @@ use super::tunnel_state::TunnelState;
 pub enum TunnelEvent {
     NewState(TunnelState),
     MixnetState(MixnetEvent),
+    ConfigChanged(Box<VpnServiceConfig>),
 }
 
 impl fmt::Display for TunnelEvent {
@@ -27,6 +30,7 @@ impl fmt::Display for TunnelEvent {
         match self {
             Self::NewState(new_state) => new_state.fmt(f),
             Self::MixnetState(event) => event.fmt(f),
+            Self::ConfigChanged(config) => config.fmt(f),
         }
     }
 }
