+++ nym-vpn-core/crates/nym-vpn-proto/src/conversions/vpnd.rs
@@ -20,7 +20,7 @@ use nym_vpnd_types::{
     AccountCommandResponse, ListCountriesOptions, ListGatewaysOptions, StoreAccountRequest,
     gateway::{Performance, Score},
     log_path::LogPath,
-    service::{ConnectArgs, ConnectOptions, VpnServiceConfig, VpnServiceInfo},
+    service::{ConnectArgs, ConnectOptions, VpnServiceInfo},
 };
 
 use crate::{conversions::ConversionError, proto};
@@ -438,8 +438,8 @@ impl TryFrom<ConnectArgs> for proto::ConnectRequest {
     type Error = ConversionError;
 
     fn try_from(value: ConnectArgs) -> Result<Self, Self::Error> {
-        let entry = value.entry.map(proto::EntryNode::try_from).transpose()?;
-        let exit = value.exit.map(proto::ExitNode::try_from).transpose()?;
+        let entry = value.entry.map(proto::EntryNode::from);
+        let exit = value.exit.map(proto::ExitNode::from);
         Ok(Self {
             dns: value.options.dns.map(|ip| proto::Dns {
                 ip: Some(ip.to_string()),
@@ -478,207 +478,6 @@ impl TryFrom<proto::ConnectRequest> for ConnectOptions {
             disable_background_cover_traffic: value.disable_background_cover_traffic,
             enable_credentials_mode: value.enable_credentials_mode,
             user_agent: value.user_agent.map(UserAgent::from),
-        })
-    }
-}
-
-impl TryFrom<proto::VpnServiceConfig> for VpnServiceConfig {
-    type Error = ConversionError;
-
-    fn try_from(value: proto::VpnServiceConfig) -> Result<Self, Self::Error> {
-        let config = VpnServiceConfig {
-            entry_point: value
-                .entry_point
-                .map(nym_gateway_directory::EntryPoint::try_from)
-                .transpose()?
-                .ok_or(ConversionError::NoValueSet("VpnServiceConfig.entry_point"))?,
-
-            exit_point: value
-                .exit_point
-                .map(nym_gateway_directory::ExitPoint::try_from)
-                .transpose()?
-                .ok_or(ConversionError::NoValueSet("VpnServiceConfig.exit_point"))?,
-
-            dns: match value.dns.and_then(|dns| dns.ip) {
-                Some(ip) => Some(
-                    ip.parse::<std::net::IpAddr>()
-                        .map_err(|e| ConversionError::ParseAddr("VpnServiceConfig.dns", e))?,
-                ),
-                None => None,
-            },
-            allow_lan: value.allow_lan,
-            disable_ipv6: value.disable_ipv6,
-            enable_two_hop: value.enable_two_hop,
-            netstack: value.netstack,
-            disable_poisson_rate: value.disable_poisson_rate,
-            disable_background_cover_traffic: value.disable_background_cover_traffic,
-            min_mixnode_performance: value.min_mixnode_performance.map(|u| u as u8),
-            min_gateway_mixnet_performance: value.min_gateway_mixnet_performance.map(|u| u as u8),
-            min_gateway_vpn_performance: value.min_gateway_vpn_performance.map(|u| u as u8),
-        };
-        Ok(config)
-    }
-}
-
-impl TryFrom<VpnServiceConfig> for proto::VpnServiceConfig {
-    type Error = ConversionError;
-
-    fn try_from(value: VpnServiceConfig) -> Result<Self, Self::Error> {
-        let config = proto::VpnServiceConfig {
-            entry_point: Some(proto::EntryNode::try_from(value.entry_point)?),
-            exit_point: Some(proto::ExitNode::try_from(value.exit_point)?),
-            dns: value.dns.map(|ip| proto::Dns {
-                ip: Some(ip.to_string()),
-            }),
-            allow_lan: value.allow_lan,
-            disable_ipv6: value.disable_ipv6,
-            enable_two_hop: value.enable_two_hop,
-            netstack: value.netstack,
-            disable_poisson_rate: value.disable_poisson_rate,
-            disable_background_cover_traffic: value.disable_background_cover_traffic,
-            min_mixnode_performance: value.min_mixnode_performance.map(|u| u as u32),
-            min_gateway_mixnet_performance: value.min_gateway_mixnet_performance.map(|u| u as u32),
-            min_gateway_vpn_performance: value.min_gateway_vpn_performance.map(|u| u as u32),
-        };
-
-        Ok(config)
-    }
-}
-
-impl TryFrom<proto::EntryNode> for nym_gateway_directory::EntryPoint {
-    type Error = ConversionError;
-
-    fn try_from(value: proto::EntryNode) -> Result<Self, Self::Error> {
-        let entry_enum_value = value
-            .entry_node_enum
-            .ok_or(ConversionError::NoValueSet("EntryNode.entry_node_enum"))?;
-
-        Ok(match entry_enum_value {
-            proto::entry_node::EntryNodeEnum::Gateway(gateway) => {
-                let identity = nym_gateway_directory::NodeIdentity::from_base58_string(&gateway.id)
-                    .map_err(|err| {
-                        ConversionError::Generic(format!("failed to parse gateway id: {err}"))
-                    })?;
-                nym_gateway_directory::EntryPoint::Gateway { identity }
-            }
-
-            proto::entry_node::EntryNodeEnum::Country(country) => {
-                nym_gateway_directory::EntryPoint::Country {
-                    two_letter_iso_country_code: country.two_letter_iso_country_code.to_string(),
-                }
-            }
-            proto::entry_node::EntryNodeEnum::Region(region) => {
-                nym_gateway_directory::EntryPoint::Region {
-                    region: region.region.to_string(),
-                }
-            }
-            proto::entry_node::EntryNodeEnum::Random(_) => {
-                nym_gateway_directory::EntryPoint::Random
-            }
-        })
-    }
-}
-
-impl TryFrom<nym_gateway_directory::EntryPoint> for proto::EntryNode {
-    type Error = ConversionError;
-    fn try_from(value: nym_gateway_directory::EntryPoint) -> Result<Self, Self::Error> {
-        match value {
-            nym_gateway_directory::EntryPoint::Gateway { identity } => Ok(proto::EntryNode {
-                entry_node_enum: Some(proto::entry_node::EntryNodeEnum::Gateway(
-                    proto::GatewayId {
-                        id: identity.to_base58_string(),
-                    },
-                )),
-            }),
-            nym_gateway_directory::EntryPoint::Country {
-                two_letter_iso_country_code,
-            } => Ok(proto::EntryNode {
-                entry_node_enum: Some(proto::entry_node::EntryNodeEnum::Country(proto::Country {
-                    two_letter_iso_country_code,
-                })),
-            }),
-            nym_gateway_directory::EntryPoint::Region { region } => Ok(proto::EntryNode {
-                entry_node_enum: Some(proto::entry_node::EntryNodeEnum::Region(proto::Region {
-                    region,
-                })),
-            }),
-            nym_gateway_directory::EntryPoint::Random => Ok(proto::EntryNode {
-                entry_node_enum: Some(proto::entry_node::EntryNodeEnum::Random(())),
-            }),
-        }
-    }
-}
-
-impl TryFrom<proto::ExitNode> for nym_gateway_directory::ExitPoint {
-    type Error = ConversionError;
-
-    fn try_from(value: proto::ExitNode) -> Result<Self, Self::Error> {
-        let exit_enum_value = value
-            .exit_node_enum
-            .ok_or(ConversionError::NoValueSet("ExitNode.exit_node_enum"))?;
-
-        Ok(match exit_enum_value {
-            proto::exit_node::ExitNodeEnum::Address(address) => {
-                let address = nym_gateway_directory::Recipient::try_from_base58_string(
-                    address.nym_address.clone(),
-                )
-                .map_err(|err| {
-                    ConversionError::Generic(format!("failed to parse exit node address: {err}"))
-                })?;
-                nym_gateway_directory::ExitPoint::Address {
-                    address: Box::new(address),
-                }
-            }
-            proto::exit_node::ExitNodeEnum::Gateway(gateway) => {
-                let identity = nym_gateway_directory::NodeIdentity::from_base58_string(&gateway.id)
-                    .map_err(|err| {
-                        ConversionError::Generic(format!("failed to parse gateway id: {err}"))
-                    })?;
-                nym_gateway_directory::ExitPoint::Gateway { identity }
-            }
-            proto::exit_node::ExitNodeEnum::Country(country) => {
-                nym_gateway_directory::ExitPoint::Country {
-                    two_letter_iso_country_code: country.two_letter_iso_country_code.to_string(),
-                }
-            }
-            proto::exit_node::ExitNodeEnum::Region(region) => {
-                nym_gateway_directory::ExitPoint::Region {
-                    region: region.region.to_string(),
-                }
-            }
-            proto::exit_node::ExitNodeEnum::Random(_) => nym_gateway_directory::ExitPoint::Random,
-        })
-    }
-}
-
-impl TryFrom<nym_gateway_directory::ExitPoint> for proto::ExitNode {
-    type Error = ConversionError;
-
-    fn try_from(value: nym_gateway_directory::ExitPoint) -> Result<Self, Self::Error> {
-        let exit_node_enum = match value {
-            nym_gateway_directory::ExitPoint::Address { address } => {
-                proto::exit_node::ExitNodeEnum::Address(proto::Address {
-                    nym_address: address.to_string(),
-                    gateway_id: address.gateway().to_base58_string(),
-                })
-            }
-            nym_gateway_directory::ExitPoint::Gateway { identity } => {
-                proto::exit_node::ExitNodeEnum::Gateway(proto::GatewayId {
-                    id: identity.to_base58_string(),
-                })
-            }
-            nym_gateway_directory::ExitPoint::Country {
-                two_letter_iso_country_code,
-            } => proto::exit_node::ExitNodeEnum::Country(proto::Country {
-                two_letter_iso_country_code,
-            }),
-            nym_gateway_directory::ExitPoint::Region { region } => {
-                proto::exit_node::ExitNodeEnum::Region(proto::Region { region })
-            }
-            nym_gateway_directory::ExitPoint::Random => proto::exit_node::ExitNodeEnum::Random(()),
-        };
-        Ok(proto::ExitNode {
-            exit_node_enum: Some(exit_node_enum),
         })
     }
 }
@@ -854,9 +653,9 @@ impl From<nym_vpnd_types::gateway::Country> for proto::Country {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::proto;
     use nym_gateway_directory::{EntryPoint, ExitPoint, NodeIdentity};
+    use nym_vpn_lib_types::service::VpnServiceConfig;
     use std::str::FromStr;
 
     #[test]
@@ -881,7 +680,7 @@ mod tests {
             min_gateway_vpn_performance: Some(1u8),
         };
 
-        let external: proto::VpnServiceConfig = internal.clone().try_into().unwrap();
+        let external: proto::VpnServiceConfig = internal.clone().into();
         let internal_converted: VpnServiceConfig = external.try_into().unwrap();
         assert_eq!(internal, internal_converted);
     }
